refactor: separate out bitcoin core utils view and start separating out quick connect.
